requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Limiter==3.3.1
Flask-Session==0.5.0
Flask-CORS==4.0.0
WTForms==3.0.1
Werkzeug==2.3.6
SQLAlchemy==2.0.18
bcrypt==4.0.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.15.0
python-dotenv==1.0.0
reportlab==4.0.4
